Refactored attendee-related files
