+++ layout/Login/components/index.ts
@@ -1,0 +1,3 @@
+import LoginForm from "./LoginForm";
+
+export { LoginForm };
