+++ layout/Profile/components/index.ts
@@ -1,0 +1,4 @@
+import CheckpointTracker from "./CheckpointTracker";
+import CodeInput from "./CodeInput";
+
+export { CheckpointTracker, CodeInput };
