+++ layout/Store/components/index.ts
@@ -1,0 +1,3 @@
+import Product from "./Product";
+
+export { Product };
