+++ layout/Vault/components/index.ts
@@ -1,0 +1,6 @@
+import Prizes from "./Prizes";
+import Redeemables from "./Redeemables";
+import StoreEmpty from "./StoreEmpty";
+import WheelEmpty from "./WheelEmpty";
+
+export { Prizes, Redeemables, StoreEmpty, WheelEmpty };
