+++ layout/Wheel/components/index.ts
@@ -1,0 +1,6 @@
+import ListItem3Cols from "./ListItem3Cols";
+import ListItem4Cols from "./ListItem4Cols";
+import WheelComponent from "./WheelComponent";
+import WheelMessage from "./WheelMessage";
+
+export { ListItem3Cols, ListItem4Cols, WheelComponent, WheelMessage };
